restructure the ProducerConf into its own file and spec
