+++ src/test/scala/com/weightwatchers/reactive/kinesis/producer/KinesisProducerActorSpec.scala
@@ -41,41 +41,6 @@ class KinesisProducerActorSpec
     with MockitoSugar
     with BeforeAndAfterAll {
 
-  val defaultKinesisConfig =
-    ConfigFactory.parseFile(new File("src/main/resources/reference.conf")).getConfig("kinesis")
-
-  val kinesisConfig = ConfigFactory
-    .parseString("""
-      |kinesis {
-      |
-      |   application-name = "TestSpec"
-      |
-      |   testProducer {
-      |      stream-name = "core-test-kinesis-producer"
-      |
-      |      akka {
-      |         max-outstanding-requests = 50000
-      |      }
-      |
-      |      kpl {
-      |         Region = us-east-1
-      |         KinesisEndpoint = "CustomKinesisEndpoint"
-      |         KinesisPort = 1111
-      |         CredentialsRefreshDelay = 5001
-      |         CloudwatchEndpoint = "CustomCloudWatchEndpoint"
-      |         CloudwatchPort = 2222
-      |         EnableCoreDumps = true
-      |         NativeExecutable = "NativeExecutable"
-      |         TempDirectory = "TempDirectory"
-      |         ThreadPoolSize = 1
-      |         ThreadingModel = "ThreadingModel.POOLED"
-      |      }
-      |   }
-      |}
-    """.stripMargin)
-    .getConfig("kinesis")
-    .withFallback(defaultKinesisConfig)
-
   implicit val timeout = Timeout(5.seconds)
 
   import system.dispatcher
@@ -86,27 +51,6 @@ class KinesisProducerActorSpec
   }
 
   "The KinesisProducerActor" - {
-
-    "Should parse the Config into a ProducerConf" in {
-      val producerConf = ProducerConf(kinesisConfig, "testProducer")
-
-      producerConf.dispatcher should be(Some("kinesis.akka.default-dispatcher"))
-      producerConf.kplConfig.getString("Region") should be("us-east-1")                            //validate an override properly
-      producerConf.kplConfig.getBoolean("AggregationEnabled") should be(true)                      //validate a default property
-      producerConf.kplConfig.getString("KinesisEndpoint") should be("CustomKinesisEndpoint")       //validate an override property
-      producerConf.kplConfig.getLong("KinesisPort") should be(1111)                                //validate an override property
-      producerConf.kplConfig.getLong("CredentialsRefreshDelay") should be(5001)                    //validate an override property
-      producerConf.kplConfig.getString("CloudwatchEndpoint") should be("CustomCloudWatchEndpoint") //validate an override property
-      producerConf.kplConfig.getLong("CloudwatchPort") should be(2222)                             //validate an override property
-      producerConf.kplConfig.getBoolean("EnableCoreDumps") should be(true)                         //validate an override property
-      producerConf.kplConfig.getString("NativeExecutable") should be("NativeExecutable")           //validate an override property
-      producerConf.kplConfig.getString("TempDirectory") should be("TempDirectory")                 //validate an override property
-      producerConf.kplConfig.getString("ThreadingModel") should be("ThreadingModel.POOLED")        //validate an override property
-      producerConf.kplConfig.getInt("ThreadPoolSize") should be(1)                                 //validate an override property
-      producerConf.throttlingConf.get.maxOutstandingRequests should be(50000)
-      producerConf.throttlingConf.get.retryDuration should be(100.millis)
-      producerConf.streamName should be("core-test-kinesis-producer")
-    }
 
     "Should process a message without a response for a Send" in {
       val probe = TestProbe()
